# Cache spec provider mods