@colemanw 2024-01-22
@eileenmcnaughton are you testing with the latest changes from https://lab.civicrm.org/dev/core/-/issues/4905 which should have dramatically sped-up the `fieldBelongsToEntity` function.