@eileenmcnaughton 2024-01-22
@colemanw I'm testing with what is merged so far.... my take is that reduced the queries to load custom data -but I'm seeing large amounts of time being used at the php layer on building the field spec - e.g 6.3 seconds spent on 130,807 calls to getCodeDocs()