@colemanw 2024-01-22
@eileenmcnaughton what's the calling function that calls `getCodeDocs` so many times? Maybe we need to use `Civi::$statics` there.