@eileenmcnaughton 2024-01-22
@colemanw it does help - but not nearly as much as calling modifySpec thousands of time less would...