@colemanw 2024-01-23
@eileenmcnaughton I think #29066 will solve it for you then, as it caches post-`SpecProvider`-mods.